Reuse bitness detection on mac/linux
